Initial, some tests (working)
- Working Makefile which produces a shared library combining Haskell code
and C lualibhelper in a single file. (Not separate linked .so)
- The library loads well into Lua and Lua is able to call Haskell
functions, which are exported via C.
- Lua opens file, passes it to C function. C function obtains a file
pointer from Lua state (first) argument and calls Haskell (C exported)
function with file pointer as argument.
- Haskell obtains FD using imported C function "fileno" and converts it
to Handle.
- Haskell function is able to write to file using this Handle. But very
important part is Handle's buffering. If not flushed before Haskell
function exits, it seems that buffered data will never be written to
disk.